+++ src/Captura.FFmpeg/FFmpegModule.cs
@@ -16,6 +16,7 @@ namespace Captura
                 Binder.Bind<IAudioWriterItem>(() => audioItem);
             }
 
+            Binder.Bind<IVideoConverter>(() => new FFmpegGifConverter());
             Binder.Bind<IVideoConverter>(() => new FFmpegVideoConverter(new Vp8VideoCodec()));
             Binder.Bind<IVideoConverter>(() => new FFmpegVideoConverter(new Vp9VideoCodec()));
         }
